Break down the project into subprojects
* Code (JSON RPC structures and converters)
* Client (Call JSON RPC services)
* Server (JSON RPC service descriptors)

Fix #11
