+++ settings.gradle.kts
@@ -1,1 +1,6 @@
+rootProject.name = "voorhees"
 rootProject.buildFileName = "build.gradle.kts"
+
+include("voorhees-core")
+include("voorhees-client")
+include("voorhees-server")
