+++ voorhees-client/src/main/kotlin/com/hylamobile/voorhees/client/annotation/JsonRpcService.kt
