+++ voorhees-core/src/main/kotlin/com/hylamobile/voorhees/jsonrpc/objects.kt
