+++ voorhees-server/src/main/kotlin/com/hylamobile/voorhees/util/Option.kt
