+++ voorhees-server/src/main/kotlin/com/hylamobile/voorhees/util/Sequence.kt
