+++ voorhees-server/src/main/resources/META-INF/spring.factories
